requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
Django==5.1.4
django-allauth==65.3.1
django-crispy-forms==2.3
sqlparse==0.5.3
mysqlclient==2.2.4
